INT-241 move replacement envelope creators
