+++ src/features/repairCode/buildReplacementEnvelope.ts
@@ -1,3 +1,4 @@
+import { createPrinter, createSourceFile, EmitHint, factory, JsxSelfClosingElement, NewLineKind, ScriptKind, ScriptTarget } from 'typescript';
 import { CaseKind } from '../../cases/types';
 import type { Classification } from '../../classifier/types';
 import type { ReplacementEnvelope } from '../../components/inferenceService';
@@ -11,12 +12,83 @@ import {
 import type { VscodeDiagnostic } from '../../vscode/types';
 import {
 	buildReplacement,
-	buildTs2741NextJsImageComponentMissingAttributeInferenceJob,
 	buildTs2769ObjectAssignReplacement,
 } from './buildReplacement';
 import { extractKindsFromTs2345ErrorMessage } from './extractKindsFromTs2345ErrorMessage';
 
-export const buildInferenceJob = (
+export const buildTs2741NextJsImageComponentMissingAttributeInferenceJob = (
+	file: File,
+	node: JsxSelfClosingElement,
+): ReplacementEnvelope => {
+	const sourceFile = createSourceFile(
+		'index.ts',
+		'',
+		ScriptTarget.Latest,
+		false,
+		ScriptKind.TS,
+	);
+
+	const printer = createPrinter({ newLine: NewLineKind.LineFeed });
+
+	const jsxAttribute = factory.createJsxAttribute(
+		factory.createIdentifier('alt'),
+		factory.createStringLiteral('', true),
+	);
+
+	const { properties } = node.attributes;
+
+	const lastAttribute = properties[properties.length - 1];
+
+	if (!lastAttribute) {
+		const jsxSelfClosingElement = factory.updateJsxSelfClosingElement(
+			node,
+			node.tagName,
+			node.typeArguments,
+			factory.createJsxAttributes(
+				[
+					jsxAttribute,
+				]
+			)
+		)
+	
+		const replacement = printer.printNode(EmitHint.Expression, jsxSelfClosingElement, sourceFile);
+
+		const start = node.getStart();
+		const end = node.getEnd();
+
+		const range = buildIntuitaRangeFromSimpleRange(
+			file.separator,
+			file.lengths,
+			{ start, end },
+		);
+
+		return {
+			range,
+			replacement,
+		};
+	}
+
+	const start = lastAttribute.getEnd();
+	const end = start;
+
+	const range = buildIntuitaRangeFromSimpleRange(
+		file.separator,
+		file.lengths,
+		{ start, end },
+	);
+
+	const width = lastAttribute.getLeadingTriviaWidth();
+	const triviaText = lastAttribute.getFullText().slice(0, width);
+
+	const replacement = printer.printNode(EmitHint.Unspecified, jsxAttribute, sourceFile);
+
+	return {
+		range,
+		replacement: `${triviaText}${replacement}`,
+	};
+}
+
+export const buildReplacementEnvelope = (
 	file: File,
 	diagnostic: VscodeDiagnostic,
 	classification: Classification,
